The true initial commit. Run src/gui/music-grid.rkt to see the gui.
lang contains new language forms and helpers.
Contains:
left-to-right-composition.rkt - a left-to-right composition form.
define-properties - a C#-like properties form for racket/class.

gui contains the graphical elements and helpers.
Contains:
music-grid.rkt - this is the application main at the moment.
grid-canvas/ - a grid canvas, which is a canvas with a grid drawn onto
it (along with helpers)
gui/ext contains racket/gui classes that extend the usual callbacks and behaviors provided.